requirements.txt
numpy
opencv-contrib-python
mysql-connector-python
pyfiglet
Pillow